Create PassHass1.php
Process to store password

To Store a Password
1.	Generate a long random salt using a CSPRNG.
2.	Prepend the salt to the password and hash it with a standard cryptographic hash function such as SHA256.
3.	Save both the salt and the hash in the user's database record.
To Validate a Password
1.	Retrieve the user's salt and hash from the database.
2.	Prepend the salt to the given password and hash it using the same hash function.
3.	Compare the hash of the given password with the hash from the database. If they match, the password is correct. Otherwise, the password is incorrect.

In theory we should use HTTPS from the browser to the server to stop the password being sniffed before it gets to the application server for processing

Our algorithm will include


Slow Hash Function and salt function
Salt ensures that attackers can't use specialized attacks like lookup tables and rainbow tables to crack large collections of hashes quickly, but it doesn't prevent them from running dictionary or brute-force attacks on each hash individually. High-end graphics cards (GPUs) and custom hardware can compute billions of hashes per second, so these attacks are still very effective. To make these attacks less effective, we can use a technique known as key stretching.
Comparing the hashes in "length-constant" time ensures that an attacker cannot extract the hash of a password in an on-line system using a timing attack, then crack it off-line.